Restored the names of the Form1 form Renamed the new application form to MainForm Begin implementing the mainform presenter
