+++ Enshrouded Server Manager/MainForm.resx
@@ -117,33 +117,6 @@
   <resheader name="writer">
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
-  <data name="lblNewsText.Text" xml:space="preserve">
-    <value>Changes:
-
-v0.4.1
-- Added option to set custom messages
-   for Discord notifications
-
-v0.4.2
-- Fixed an issue with server names not
-   allowing spaces
-- Fixed an issue with button visibility
-   when updating a server</value>
-  </data>
-  <data name="label1.Text" xml:space="preserve">
-    <value>The Enshrouded server saves
-backups every 5 minutes, but
-it overwrites the previous file.
-
-This feature will save all the
-server files in the save directory
-at a configured interval. When
-the maximum backups is reached
-the oldest backup is deleted.
-
-Auto Backup will only run while
-Enshrouded Server Manager is open.</value>
-  </data>
   <assembly alias="System.Drawing" name="System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a" />
   <data name="$this.Icon" type="System.Drawing.Icon, System.Drawing" mimetype="application/x-microsoft.net.object.bytearray.base64">
     <value>
