+++ Enshrouded Server Manager/Views/Interfaces/IMainFormView.cs
@@ -1,0 +1,6 @@
+﻿namespace Enshrouded_Server_Manager.Views;
+public interface IMainFormView
+{
+    event EventHandler ViewCreditsButtonClicked;
+    string CurrentVersionText { get; set; }
+}
